requirements2.txt
numpy==1.26.4
streamlit==1.20.0
openai==0.28.1
jsonschema==4.19.0
textblob==0.17.1
matplotlib==3.8.0
weaviate-client==3.6.0

